+++ CHANGELOG.md
@@ -14,12 +14,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Automatically determine the number of layers
 - Add a setting to enable frame previews in silhouette mode
 - Add non square pixel size support
 - Add `orfail` crate to the dependencies
 
 ### Removed
 
+- Remove layer count setting
 - Remove max undos setting
 - Remove import-from-clipboard feature
 - Remove finger mode